pyproject.toml
[project]
name = "genx-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-eventgrid>=4.22.0",
    "azure-functions>=1.23.0",
    "azure-identity>=1.23.0",
    "azure-mgmt-containerinstance>=10.1.0",
    "azure-storage-blob>=12.25.1",
    "black[jupyter]>=25.1.0",
    "joblib>=1.5.1",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "nbstripout>=0.8.1",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-mock>=3.14.1",
    "requests>=2.32.3",
    "universal-pathlib>=0.2.6",
    "xlwings>=0.33.14",
]
